% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmers}
\alias{kmers}
\title{kmers}
\usage{
kmers(s, bin = 3L, step = 1L)
}
\arguments{
\item{s}{seq}

\item{bin}{kmer size}

\item{step}{kmer step}
}
\description{
kmers
}
\examples{
kmers("AGCTTTTTTTTT")
kmers("AGCTTTTTTTTT", 4)

}
