% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmers_tm}
\alias{kmers_tm}
\title{get kmers by tm range}
\usage{
kmers_tm(s, t0, dt = 1)
}
\arguments{
\item{s}{seq}

\item{t0}{kmer tm}

\item{dt}{kmer tm range}
}
\description{
get kmers by tm range
}
\examples{
kmers_tm("AGCTTTTTTTTT", 8)
}
