% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev}
\alias{rev}
\title{Reverse a Vector of Strings}
\usage{
rev(x)
}
\arguments{
\item{x}{a character vector.}
}
\description{
Reverses a vector of 'strings' (a character vector). Not safe for
unicode (UTF-8) characters.
}
\examples{
x <- c("ABC", "DEF", "GHIJ")
rev(x)
}
